# Conflicts:
#	.gitignore

+++ src/org/jitsi/impl/neomedia/transform/srtp/SRTCPCryptoContext.java
@@ -19,7 +19,7 @@ import java.util.*;
 
 import org.bouncycastle.crypto.params.*;
 import org.jitsi.bccontrib.params.*;
-import org.jitsi.impl.neomedia.*;
+import org.jitsi.service.neomedia.*;
 
 /**
  * SRTPCryptoContext class is the core class of SRTP implementation. There can
@@ -211,7 +211,7 @@ public class SRTCPCryptoContext
      */
     public void processPacketAESCM(RawPacket pkt, int index)
     {
-        int ssrc = pkt.getRTCPSSRC();
+        int ssrc = (int) pkt.getRTCPSSRC();
 
         /* Compute the CM IV (refer to chapter 4.1.1 in RFC 3711):
         *
